made test broadcast notification

+++ core/src/main/java/io/github/_127_0_0_l/core/models/Notification.java
@@ -1,6 +1,6 @@
 package io.github._127_0_0_l.core.models;
 
 public record Notification(
-        int chatId,
+        Long chatId,
         VehicleAdvert vehicleAdvert
 ) { }
